+++ src/main/java/exceptions/NotFoundException.java
@@ -1,7 +1,7 @@
 package exceptions;
 
 public class NotFoundException extends RuntimeException {
-    public NotFoundException(int id) {
+    public NotFoundException(long id) {
         super("Il record con id"+id+"non è stato trovato");
     }
 }
